# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Layal Al Zahra Salon CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "401":
          description: Not authenticated

  # STAFF
...
  /appointments/summary:
    get:
      operationId: getAppointmentSummary
      tags: [appointments]
      summary: Get appointment summary analytics for a date range
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Appointment summary analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentSummary"
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: staffId
          in: query
          schema:
            type: integer
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: serviceId
          in: query
          schema:
            type: integer
        - name: paymentMethod
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete/cancel an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /appointments/{id}/status:
    patch:
      operationId: updateAppointmentStatus
      tags: [appointments]
      summary: Update appointment status (confirm, complete, no-show, cancel)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  # SALES
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/StaffMember"

    # Staff
    StaffMember:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, receptionist, stylist, therapist]
        specialization:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        commissionRate:
          type: ["number", "null"]
        targetMonthly:
          type: ["number", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Appointment:
      type: object
      required: [id, clientId, clientName, staffId, staffName, serviceId, serviceName, date, startTime, endTime, status, totalPrice, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: ["string", "null"]
        staffId:
          type: integer
        staffName:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        serviceCategoryName:
          type: ["string", "null"]
        serviceDuration:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled, no_show]
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
        staffColor:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        saleId:
          type: ["integer", "null"]
        createdAt:
          type: string
    AppointmentSummary:
      type: object
      required: [totals, byStatus, byDayHour, byStaff, byCategory, topServicesByCount, topServicesByRevenue, topStaffByCompleted]
      properties:
        totals:
          type: object
          required: [total, completed, cancelled, noShow, pending, confirmed, inProgress, totalRevenue, avgRevenue, newClients, returningClients]
          properties:
            total:
              type: integer
            completed:
              type: integer
            cancelled:
              type: integer
            noShow:
              type: integer
            pending:
              type: integer
            confirmed:
              type: integer
            inProgress:
              type: integer
            totalRevenue:
              type: number
            avgRevenue:
              type: number
            newClients:
              type: integer
            returningClients:
              type: integer
        byStatus:
          type: array
          items:
            type: object
            required: [status, count]
            properties:
              status:
                type: string
              count:
                type: integer
        byDayHour:
          type: array
          items:
            type: object
            required: [dayOfWeek, hour, count]
            properties:
              dayOfWeek:
                type: integer
              hour:
                type: integer
              count:
                type: integer
        byStaff:
          type: array
          items:
            type: object
            required: [staffId, staffName, count, completed, revenue]
            properties:
              staffId:
                type: integer
              staffName:
                type: string
              count:
                type: integer
              completed:
                type: integer
              revenue:
                type: number
        byCategory:
          type: array
          items:
            type: object
            required: [categoryName, count]
            properties:
              categoryName:
                type: string
              count:
                type: integer
        topServicesByCount:
          type: array
          items:
            type: object
            required: [serviceId, serviceName, count, revenue]
            properties:
              serviceId:
                type: integer
              serviceName:
                type: string
              count:
                type: integer
              revenue:
                type: number
        topServicesByRevenue:
          type: array
          items:
            type: object
            required: [serviceId, serviceName, count, revenue]
            properties:
              serviceId:
                type: integer
              serviceName:
                type: string
              count:
                type: integer
              revenue:
                type: number
        topStaffByCompleted:
          type: array
          items:
            type: object
            required: [staffId, staffName, completed, revenue]
            properties:
              staffId:
                type: integer
              staffName:
                type: string
              completed:
                type: integer
              revenue:
                type: number
    AppointmentInput:
      type: object
      required: [clientId, staffId, serviceId, date, startTime]
      properties:
        clientId:
          type: integer
        staffId:
          type: integer
        serviceId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        notes:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        clientId:
          type: integer
        staffId:
          type: integer
        serviceId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        notes:
          type: string
    AppointmentStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled, no_show]

    # Sales
...